took out INT function in int(4) to int(2) conversions. Incorporated nick's ie regent scale change. reverted one change I made in intree.
git-svn-id: svn+ssh://svn.code.sf.net/p/open-fvs/code/branches/GaryDev@2122 90b89b48-9230-0de4-5065-812dd12741b8

+++ dbs/src/dbscmpu.f
@@ -345,7 +345,7 @@ C
 
       DO X=1,(NUMCMPU)
         CURVAL(X)=KWVALS(X)
-        ColNumber=INT(ColNumber+1)
+        ColNumber=ColNumber+1
         iRet = fvsSQLBindParameter(StmtHndlOut,ColNumber,
      -           SQL_PARAM_INPUT,
      -           SQL_F_DOUBLE, SQL_DOUBLE,INT(15,SQLUINTEGER_KIND),
